+++ core/CHANGELOG.md
@@ -25,6 +25,8 @@
 
 - Remove `EitherUpgrade` in favor of implementing `UpgradeInfo`, `InboundUpgrade` and `OutboundUpgrade` on `either::Either`. See [PR 3339].
 
+- Remove `EitherFuture2` in favor of `EitherFuture`. See [PR 3340].
+
 [PR 3031]: https://github.com/libp2p/rust-libp2p/pull/3031
 [PR 3058]: https://github.com/libp2p/rust-libp2p/pull/3058
 [PR 3097]: https://github.com/libp2p/rust-libp2p/pull/3097
@@ -33,6 +35,7 @@
 [PR 3337]: https://github.com/libp2p/rust-libp2p/pull/3337
 [PR 3338]: https://github.com/libp2p/rust-libp2p/pull/3338
 [PR 3339]: https://github.com/libp2p/rust-libp2p/pull/3339
+[PR 3340]: https://github.com/libp2p/rust-libp2p/pull/3340
 
 # 0.37.0